# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: CasaRoca System AI · API
  version: "1.0.0"
  description: |
    API del sistema eclesial de Casa Sobre la Roca: una central y 36 sedes.

    ESTE ARCHIVO SE GENERA. No lo edite a mano: corra `scripts/generar-openapi.js`.
    La verificacion de entrega compara lo generado con lo versionado.

    Autenticacion: token de portador (Bearer). El token dice que sesion DICE ser;
    quien decide es la base, que en cada peticion responde si esa sesion sigue viva.
    Por eso cerrar una sesion surte efecto en el instante.
servers:
  - url: https://api.casaroca.org
    description: produccion
  - url: https://api.staging.casaroca.org
    description: pruebas
  - url: http://127.0.0.1:3000
    description: desarrollo
components:
  securitySchemes:
    portador:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error: { type: boolean }
        mensaje: { type: string, description: "En castellano y con la accion siguiente. Nunca filtra la estructura interna." }
        peticionId: { type: string, description: "Identificador de traza. El usuario lo lee por telefono al soporte." }
security:
  - portador: []
paths:
...
  /api/v1/consejeria/casos/:id/asignar:
    post:
      summary: "Asignar un consejero"
      description: "El caso pasa a «en proceso». Un caso sin nadie detras es una persona esperando."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/consejeria/casos/:id/cerrar:
    post:
      summary: "Cerrar o derivar"
      description: "Cerrar EXIGE `desenlace` (como termino). Derivar EXIGE `derivadoA`. Cerrar el caso cierra tambien las asignaciones vivas."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/consejeria/casos/:id/notas:
    post:
      summary: "Escribir una nota"
      description: "Va aparte a proposito: no viaja en ninguna lista, solo se ve dentro de la ficha."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/consejeria/casos/:id/sesiones:
    post:
      summary: "Registrar una sesion"
      description: "Fecha, duracion, modalidad y si la persona asistio."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/consejeria/casos/:id:
    get:
      summary: "La ficha del caso"
      description: "Aqui SI van las notas, y entrar deja rastro en la bitacora de lectura, con nombre y hora. Saberlo es la mitad de la proteccion."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/consejeria/casos:
    get:
      summary: "La bandeja de casos"
      description: "⛔ SIN notas y SIN el detalle: solo lo que permite priorizar. Una lista que trae las notas de paso las reparte por accidente."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
    post:
      summary: "Abrir un caso"
      description: "Consultante y topico; la sede es la del consultante salvo que se diga otra. Antes daba 403 siempre por un RETURNING bajo la politica de lectura; ahora se inserta y se lee aparte."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
  /api/v1/consejeria/topicos:
    get:
      summary: "Los topicos"
      description: "Para que nadie escriba «problema familiar» de doce formas distintas. Marca los que exigen profesional."
      responses:
        "200": { description: correcto }
        "401": { description: "sin sesion valida", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "429": { description: "demasiadas peticiones" }
        "500": { description: "error inesperado", content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...